requirements.txt
autopep8
django~=3.2.4
django-rest-framework
django-cors-headers
djangorestframework~=3.12.4